590 project part 1 (user input) code
I've been working on the user input part of our project and this is what I came up with. It accepts user input data and has a drop-down option for default datasets. let me know what you think.

I will probably include a screenshot of the output of this in our project report due 
Monday.